Separate toPrepare button from toGrade button
