+++ src/main/java/homeschooltracker/gui/JButtonParentToGradeTask.java
@@ -1,21 +1,21 @@
-package homeschooltracker;
+package homeschooltracker.gui;
 
 import homeschooltracker.content.Task;
 
 import javax.swing.*;
 import java.awt.*;
 
-public class JButtonParentTask extends JButton {
+public class JButtonParentToGradeTask extends JButton {
     Task task;
 
-    public JButtonParentTask(String label, Task task) {
+    public JButtonParentToGradeTask(String label, Task task) {
         super(label);
         this.task = task;
     }
 
     public void execute() {
-        System.out.println("execute JButtonParentTask: " + task.getName());
-        task.setPrepared();
+        System.out.println("execute JButtonParentToGradeTask: " + task.getName());
+        task.setGraded();
         System.out.println("  --> " + task.getStateDescription());
 
         // Remove the button once completed!
